Some changes in project structure
