+++ VictoryCenter/VictoryCenter.BLL/Commands/Categories/Delete/DeleteCategoryCommand.cs
