+++ VictoryCenter/VictoryCenter.BLL/Commands/Categories/Delete/DeleteCategoryHandler.cs
