+++ VictoryCenter/VictoryCenter.BLL/Validators/Categories/CreateCategoryValidator.cs
@@ -1,15 +1,12 @@
 using FluentValidation;
+using VictoryCenter.BLL.Commands.Categories.CreateCategory;
 
-namespace VictoryCenter.BLL.Commands.Categories.CreateCategory;
+namespace VictoryCenter.BLL.Validators.Categories;
 
 public class CreateCategoryValidator : AbstractValidator<CreateCategoryCommand>
 {
     public CreateCategoryValidator()
     {
-        RuleFor(command => command.createCategoryDto.Name)
-            .NotNull()
-            .WithMessage("Name is required");
-        
         RuleFor(command => command.createCategoryDto.Name)
             .NotEmpty()
             .WithMessage("Name can't be empty");
