+++ VictoryCenter/VictoryCenter.IntegrationTests/ControllerTests/Categories/Delete/DeleteCategoryTests.cs
@@ -1,0 +1,38 @@
+using Microsoft.EntityFrameworkCore;
+using VictoryCenter.DAL.Data;
+using VictoryCenter.IntegrationTests.ControllerTests.Base;
+
+namespace VictoryCenter.IntegrationTests.ControllerTests.Categories.Delete;
+
+[Collection("SharedIntegrationTests")]
+public class DeleteCategoryTests
+{
+    private readonly HttpClient _httpClient;
+    private readonly VictoryCenterDbContext _dbContext;
+
+    public DeleteCategoryTests(IntegrationTestDbFixture fixture)
+    {
+        _httpClient = fixture.HttpClient;
+        _dbContext = fixture.DbContext;
+    }
+    
+    [Fact]
+    public async Task DeleteCategory_ShouldDeleteCategory()
+    {
+        var existingEntity = await _dbContext.Categories.FirstOrDefaultAsync();
+        
+        var response = await _httpClient.DeleteAsync($"api/categories/deleteCategory/{existingEntity!.Id}");
+        
+        Assert.True(response.IsSuccessStatusCode);
+    }
+
+    [Theory]
+    [InlineData(-1)]
+    [InlineData(0)]
+    public async Task DeleteCategory_ShouldNotDeleteCategory_NotFound(int testId)
+    {
+        var response = await _httpClient.DeleteAsync($"api/categories/deleteCategory/{testId}");
+        
+        Assert.False(response.IsSuccessStatusCode);
+    }
+}
