+++ VictoryCenter/VictoryCenter.UnitTests/MediatRHandlersTests/Categories/CreateCategoryTests.cs
@@ -3,6 +3,7 @@ using FluentValidation;
 using Moq;
 using VictoryCenter.BLL.Commands.Categories.CreateCategory;
 using VictoryCenter.BLL.DTOs.Categories;
+using VictoryCenter.BLL.Validators.Categories;
 using VictoryCenter.DAL.Entities;
 using VictoryCenter.DAL.Repositories.Interfaces.Base;
