+++ VictoryCenter/VictoryCenter.WebAPI/Controllers/BaseApiController.cs
@@ -1,6 +1,5 @@
 using FluentResults;
 using MediatR;
-using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 
 namespace VictoryCenter.WebAPI.Controllers;
